+++ apps/soybean-admin/src/typings/elegant-router.d.ts
@@ -27,6 +27,7 @@ declare module "@elegant-router/types" {
     "games_member-detail": "/games/member-detail/:id";
     "games_members": "/games/members";
     "games_play-attempts": "/games/play-attempts";
+    "games_prize-types": "/games/prize-types";
     "games_prizes": "/games/prizes";
     "games_scores": "/games/scores";
     "home": "/home";
@@ -108,6 +109,7 @@ declare module "@elegant-router/types" {
     | "games_member-detail"
     | "games_members"
     | "games_play-attempts"
+    | "games_prize-types"
     | "games_prizes"
     | "games_scores"
     | "home"